Fix silly mistake (not checking the caching option) in previous commit...
git-svn-id: http://fofix.googlecode.com/svn/MFH-Mod/trunk@1176 48f48816-a574-47fe-9a59-6914b3aef992

+++ src/Song.py
@@ -3743,7 +3743,7 @@ def getAvailableSongs(engine, library = DEFAULT_LIBRARY, includeTutorials = Fals
   for name in names:
     progressCallback(len(songs)/float(len(names)))
     songs.append(SongInfo(engine.resource.fileName(library, name, "song.ini", writable = True), library, allowCacheUsage=True))
-  if len(songs) and canCache:
+  if len(songs) and canCache and Config.get("performance", "cache_song_metadata"):
     cache = cacheManager.getCache(songs[0].fileName)
     #stump: clean up the cache
     if cache.execute('DELETE FROM `songinfo` WHERE `hash` NOT IN (' + ','.join("'%s'" % s.stateHash for s in songs) + ')').rowcount > 0:
